Move Shared Math Utils to Utils
A test fixture, 'math/Constants.tests.js' wasnt a test file but a set of fixures used by:

animation/AnimationMixer.tests.js
core/BufferGeometry.tests.js
core/Object3D.tests.js
core/Uniform.tests.js
math/Box2.tests.js
math/Box3.tests.js
math/Color.tests.js
math/Cylindrical.tests.js
math/Euler.tests.js
math/Frustum.tests.js
math/Line3.tests.js
math/Matrix4.tests.js
math/Plane.tests.js
math/Quaternion.tests.js
math/Ray.tests.js
math/Sphere.tests.js
math/Spherical.tests.js
math/Triangle.tests.js
math/Vector2.tests.js
math/Vector3.tests.js
math/Vector4.tests.js

Shared math constants for tests moved to utils.  Only tests in test source.
