+++ test/unit/src/math/Frustum.tests.js
@@ -9,7 +9,7 @@ import { Matrix4 } from '../../../../src/math/Matrix4.js';
 import { Box3 } from '../../../../src/math/Box3.js';
 import { Mesh } from '../../../../src/objects/Mesh.js';
 import { BoxGeometry } from '../../../../src/geometries/BoxGeometry.js';
-import { zero3, one3, eps } from './Constants.tests.js';
+import { zero3, one3, eps } from '../../utils/math-constants.js';
 
 const unit3 = new Vector3( 1, 0, 0 );
