+++ test/unit/src/math/Vector2.tests.js
@@ -7,7 +7,7 @@ import {
 	x,
 	y,
 	eps
-} from './Constants.tests.js';
+} from '../../utils/math-constants.js';
 
 export default QUnit.module( 'Maths', () => {
