+++ test/unit/src/math/Vector4.tests.js
@@ -9,7 +9,7 @@ import {
 	z,
 	w,
 	eps
-} from './Constants.tests.js';
+} from '../../utils/math-constants.js';
 
 export default QUnit.module( 'Maths', () => {
